# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Princess Empire E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: sort
          in: query
          schema:
            type: ["string", "null"]
            enum: [newest, price_asc, price_desc, popular, null]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: bestSeller
          in: query
          schema:
            type: ["boolean", "null"]
        - name: flashSale
          in: query
          schema:
            type: ["boolean", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: brand
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsPage"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/recent:
    get:
      operationId: getRecentProducts
      tags: [products]
      summary: Get recently uploaded products
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Recent products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/bestsellers:
    get:
      operationId: getBestSellers
      tags: [products]
      summary: Get best-selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Best sellers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/flashsales:
    get:
      operationId: getFlashSales
      tags: [products]
      summary: Get flash sale products
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Flash sale products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/search-suggestions:
    get:
      operationId: getSearchSuggestions
      tags: [products]
      summary: Get search suggestions (autocomplete)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Product:
      type: object
      required: [id, name, price, images, categoryId, stock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        discount:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        brand:
          type: ["string", "null"]
        featured:
          type: boolean
        bestSeller:
          type: boolean
        flashSale:
          type: boolean
        flashSaleEndsAt:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, images, stock]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        discount:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        brand:
          type: ["string", "null"]
        featured:
          type: boolean
        bestSeller:
          type: boolean
        flashSale:
          type: boolean
        flashSaleEndsAt:
          type: ["string", "null"]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        oldPrice:
          type: ["number", "null"]
        discount:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        brand:
          type: ["string", "null"]
        featured:
          type: boolean
        bestSeller:
          type: boolean
        flashSale:
          type: boolean
        flashSaleEndsAt:
          type: ["string", "null"]
    ProductsPage:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...