# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Princess Empire E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  images: string[];
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName?: string | null;
  sizes?: string[];
  colors?: string[];
  stock: number;
  /** @nullable */
  brand?: string | null;
  featured?: boolean;
  bestSeller?: boolean;
  flashSale?: boolean;
  /** @nullable */
  flashSaleEndsAt?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  createdAt: string;
}